finished questions 2 and 3
The puzzles are now encrypted and are able to be written to a file.

But the function `FileOutputStream.write(byte[])` seems to have an unreliable behaviour. Upon hexdump of the generated binary files, I could notice their length was not adequate.

We will have to investigate this further, but other than that, everything works.